requirements.txt
Flask~=3.1.2
Flask-SQLAlchemy
SQLAlchemy~=2.0.44
python-dotenv~=1.2.1
Werkzeug~=3.1.3
PyGObject>=3.30.0
pywebview~=5.3
nmcli
pydbus~=0.6.0